refactor(ClientEncryptor.cs): refactor DecryptPacket and EncryptPacket methods to improve code readability and maintainability feat(ClientEncryptor.cs): add CalculateKeys method overload to accept individual version components for easier key calculation feat(ClientEncryptor.cs): add CalculateKeys method to calculate client keys based on version info feat(ClientEncryptor.cs): add legacy Decrypt and Encrypt methods for backward compatibility with byte arrays feat(ClientEncryptor.cs): add helper method CalculateKeys to calculate keys from ClientVersion instance feat(ClientEncryptor.cs): change ClientEncryptor class to public to allow for inheritance and extension feat(ClientKeys.cs): change Key1 and Key2 properties in ClientKeys struct to uint for consistency and efficiency

+++ src/Moongate.UO/Encryption/ClientKeys.cs
@@ -5,6 +5,6 @@ namespace Moongate.UO.Encryption;
 /// </summary>
 public struct ClientKeys
 {
-    public long Key1 { get; set; }
-    public long Key2 { get; set; }
+    public uint Key1 { get; set; }
+    public uint Key2 { get; set; }
 }
